+++ lib/domain/initial_setting/pill_type/initial_setting_pill_category_type_page.dart
